+++ stapi-pydantic/src/stapi_pydantic/__init__.py
@@ -1,6 +1,7 @@
 from .conformance import Conformance
 from .constraints import Constraints
 from .datetime_interval import DatetimeInterval
+from .filter import CQL2Filter
 from .json_schema_model import JsonSchemaModel
 from .opportunity import (
     Opportunity,
@@ -31,6 +32,7 @@ from .shared import Link
 __all__ = [
     "Conformance",
     "Constraints",
+    "CQL2Filter",
     "DatetimeInterval",
     "JsonSchemaModel",
     "Link",